Extract router configuration from matsMethods
matsMethods was growing quite large. This change pulls the various parts
of our router configuration out of matsMethods into a number of new
modules. The hope is that doing so will make this easier to maintain.

- routeConfig - is where routes paths are definied and configured
- routeHandlers - are where the handlers are defined
- handlerHelpers - are helper functions for the route handlers
- routeRegistration - configures and sets up the router
- matsMethods - contains functions used by other parts of MATS
- matsHelpers - contains functions used by matsMethods and
  routerHandlers

In theory, matsHelpers could be further refactored.
