+++ meteor_packages/mats-common/imports/startup/api/matsMethods.js
@@ -2,23 +2,27 @@
  * Copyright (c) 2021 Colorado State University and Regents of the University of Colorado. All rights reserved.
  */
 
+/**
+ * Internal data functions for MATS
+ * Also handles route registration
+ */
+
 import { Meteor } from "meteor/meteor";
-import { WebApp } from "meteor/webapp";
 import { ValidatedMethod } from "meteor/mdg:validated-method";
 import SimpleSchema from "meteor/aldeed:simple-schema";
 import {
   matsCache,
   matsCollections,
-  matsDataQueryUtils,
   matsCouchbaseUtils,
   matsDataUtils,
   matsTypes,
   versionInfo,
 } from "meteor/randyp:mats-common";
 import { moment } from "meteor/momentjs:moment";
-import { _ } from "meteor/underscore";
 import { Mongo } from "meteor/mongo";
 import { curveParamsByApp } from "../both/mats-curve-params";
+import { setupRouter } from "./routeRegistration";
+import { checkMetaDataRefresh, getFlattenedResultData } from "./matsHelpers";
 
 /* eslint-disable no-await-in-loop */
 /* eslint-disable no-console */
@@ -39,129 +43,10 @@ const isEmpty = function (map) {
   return mapKeys.length === 0;
 };
 
-// private middleware for determining if the app is running
-const getHealth = async function (res) {
-  if (Meteor.isServer) {
-    res.sendStatus(200);
-  }
-};
-
 // wrapper for NaN check
 const isThisANaN = function (val) {
   // eslint-disable-next-line no-restricted-globals
   return val === undefined || val === null || isNaN(val);
-};
-
-// private - used to see if the main page needs to update its selectors
-const checkMetaDataRefresh = async function () {
-  // This routine compares the current last modified time of the tables (MYSQL) or documents (Couchbase)
-  // used for curveParameter metadata with the last update time to determine if an update is necessary.
-  // We really only do this for Curveparams
-  /*
-        metaDataTableUpdates:
-        {
-            name: dataBaseName(MYSQL) or bucketName(couchbase),
-            (for couchbase tables are documents)
-            tables: [tableName1, tableName2 ..],
-            lastRefreshed : timestamp
-        }
-     */
-  let refresh = false;
-  const tableUpdates = await metaDataTableUpdates.find({}).fetchAsync();
-  const settings = await matsCollections.Settings.findOneAsync();
-  const dbType = settings !== undefined ? settings.dbType : matsTypes.DbTypes.mysql;
-  for (let tui = 0; tui < tableUpdates.length; tui += 1) {
-    const id = tableUpdates[tui]._id;
-    const poolName = tableUpdates[tui].pool;
-    const dbName = tableUpdates[tui].name;
-    const tableNames = tableUpdates[tui].tables;
-    const { lastRefreshed } = tableUpdates[tui];
-    let updatedEpoch = Number.MAX_VALUE;
-    let rows;
-    let doc;
-    for (let ti = 0; ti < tableNames.length; ti += 1) {
-      const tName = tableNames[ti];
-      try {
-        if (Meteor.isServer) {
-          switch (dbType) {
-            case matsTypes.DbTypes.mysql:
-              rows = await matsDataQueryUtils.queryMySQL(
-                global[poolName],
-                `SELECT UNIX_TIMESTAMP(UPDATE_TIME)` +
-                  `    FROM   information_schema.tables` +
-                  `    WHERE  TABLE_SCHEMA = '${dbName}'` +
-                  `    AND TABLE_NAME = '${tName}'`
-              );
-              updatedEpoch = rows[0]["UNIX_TIMESTAMP(UPDATE_TIME)"];
-              break;
-            case matsTypes.DbTypes.couchbase:
-              // the tName for couchbase is supposed to be the document id
-              doc = await global.cbPool.getCB(tName);
-              updatedEpoch = doc.updated;
-              break;
-            default:
-              throw new Meteor.Error("resetApp: undefined DbType");
-          }
-        }
-        // console.log("DB says metadata for table " + dbName + "." + tName + " was updated at " + updatedEpoch);
-        if (
-          updatedEpoch === undefined ||
-          updatedEpoch === null ||
-          updatedEpoch === "NULL" ||
-          updatedEpoch === Number.MAX_VALUE
-        ) {
-          // if time of last update isn't stored by the database (thanks, Aurora DB), refresh automatically
-          // console.log("_checkMetaDataRefresh - cannot find last update time for database: " + dbName + " and table: " + tName);
-          refresh = true;
-          // console.log("FORCED Refreshing the metadata for table because updatedEpoch is undefined" + dbName + "." + tName + " : updated at " + updatedEpoch);
-          break;
-        }
-      } catch (e) {
-        throw new Error(
-          `_checkMetaDataRefresh - error finding last update time for database: ${dbName} and table: ${tName}, ERROR:${e.message}`
-        );
-      }
-      const lastRefreshedEpoch = moment.utc(lastRefreshed).valueOf() / 1000;
-      const updatedEpochMoment = moment.utc(updatedEpoch).valueOf();
-      // console.log("Epoch of when this app last refreshed metadata for table " + dbName + "." + tName + " is " + lastRefreshedEpoch);
-      // console.log("Epoch of when the DB says table " + dbName + "." + tName + " was last updated is " + updatedEpochMoment);
-      if (lastRefreshedEpoch < updatedEpochMoment || updatedEpochMoment === 0) {
-        // Aurora DB sometimes returns a 0 for last updated. In that case, do refresh the metadata.
-        refresh = true;
-        // console.log("Refreshing the metadata in the app selectors because table " + dbName + "." + tName + " was updated at " + moment.utc(updatedEpoch * 1000).format("YYYY-MM-DD HH:mm:ss") + " while the metadata was last refreshed at " + moment.utc(lastRefreshedEpoch * 1000).format("YYYY-MM-DD HH:mm:ss"));
-        break;
-      } else {
-        // console.log("NOT Refreshing the metadata for table " + dbName + "." + tName + " : updated at " + moment.utc(updatedEpoch * 1000).format("YYYY-MM-DD HH:mm:ss") + " : metadata last refreshed at " + moment.utc(lastRefreshedEpoch * 1000).format("YYYY-MM-DD HH:mm:ss"));
-      }
-    }
-    if (refresh === true) {
-      // refresh the app metadata
-      // app specific routines
-      for (let ai = 0; ai < global.appSpecificResetRoutines.length; ai += 1) {
-        await global.appSpecificResetRoutines[ai]();
-      }
-      // remember that we updated ALL the metadata tables just now
-      await metaDataTableUpdates.updateAsync(
-        {
-          _id: id,
-        },
-        {
-          $set: {
-            lastRefreshed: moment().format(),
-          },
-        }
-      );
-    }
-  }
-  return true;
-};
-
-// private middleware for clearing the cache
-const clearCache = function (res) {
-  if (Meteor.isServer) {
-    matsCache.clear();
-    res.end("<body><h1>clearCache Done!</h1></body>");
-  }
 };
 
 // private middleware for dropping a distinct instance (a single run) of a scorecard
@@ -191,1380 +76,6 @@ const dropThisScorecardInstance = async function (
     return {
       error: err.message,
     };
-  }
-};
-
-// helper function for returning an array of database-distinct apps contained within a larger MATS app
-async function getListOfApps() {
-  let apps;
-  if (
-    matsCollections.database !== undefined &&
-    (await matsCollections.database.findOneAsync({ name: "database" })) !== undefined
-  ) {
-    // get list of databases (one per app)
-    apps = (
-      await matsCollections.database.findOneAsync({
-        name: "database",
-      })
-    ).options;
-    if (!Array.isArray(apps)) apps = Object.keys(apps);
-  } else if (
-    matsCollections.variable !== undefined &&
-    (await matsCollections.variable.findOneAsync({
-      name: "variable",
-    })) !== undefined &&
-    matsCollections.threshold !== undefined &&
-    (await matsCollections.threshold.findOneAsync({
-      name: "threshold",
-    })) !== undefined
-  ) {
-    // get list of apps (variables in apps that also have thresholds)
-    apps = (
-      await matsCollections.variable.findOneAsync({
-        name: "variable",
-      })
-    ).options;
-    if (!Array.isArray(apps)) apps = Object.keys(apps);
-  } else {
-    apps = [(await matsCollections.Settings.findOneAsync()).Title];
-  }
-  return apps;
-}
-
-// helper function to map a results array to specific apps
-async function mapArrayToApps(result) {
-  // put results in a map keyed by app
-  const newResult = {};
-  const apps = await getListOfApps();
-  for (let aidx = 0; aidx < apps.length; aidx += 1) {
-    if (result[aidx] === apps[aidx]) {
-      newResult[apps[aidx]] = [result[aidx]];
-    } else {
-      newResult[apps[aidx]] = result;
-    }
-  }
-  return newResult;
-}
-
-// helper function to map a results map to specific apps
-async function mapMapToApps(result) {
-  // put results in a map keyed by app
-  let newResult = {};
-  let tempResult;
-  const apps = await getListOfApps();
-  const resultKeys = Object.keys(result);
-  if (!matsDataUtils.arraysEqual(apps.sort(), resultKeys.sort())) {
-    if (resultKeys.includes("Predefined region")) {
-      tempResult = result["Predefined region"];
-    } else {
-      tempResult = result;
-    }
-    for (let aidx = 0; aidx < apps.length; aidx += 1) {
-      newResult[apps[aidx]] = tempResult;
-    }
-  } else {
-    newResult = result;
-  }
-  return newResult;
-}
-
-// helper function for returning a map of database-distinct apps contained within a larger MATS app and their DBs
-async function getListOfAppDBs() {
-  let apps;
-  const result = {};
-  let aidx;
-  if (
-    matsCollections.database !== undefined &&
-    (await matsCollections.database.findOneAsync({ name: "database" })) !== undefined
-  ) {
-    // get list of databases (one per app)
-    apps = (
-      await matsCollections.database.findOneAsync({
-        name: "database",
-      })
-    ).options;
-    if (!Array.isArray(apps)) apps = Object.keys(apps);
-    for (aidx = 0; aidx < apps.length; aidx += 1) {
-      result[apps[aidx]] = (
-        await matsCollections.database.findOneAsync({
-          name: "database",
-        })
-      ).optionsMap[apps[aidx]].sumsDB;
-    }
-  } else if (
-    matsCollections.variable !== undefined &&
-    (await matsCollections.variable.findOneAsync({
-      name: "variable",
-    })) !== undefined &&
-    matsCollections.threshold !== undefined &&
-    (await matsCollections.threshold.findOneAsync({
-      name: "threshold",
-    })) !== undefined
-  ) {
-    // get list of apps (variables in apps that also have thresholds)
-    apps = (
-      await matsCollections.variable.findOneAsync({
-        name: "variable",
-      })
-    ).options;
-    if (!Array.isArray(apps)) apps = Object.keys(apps);
-    for (aidx = 0; aidx < apps.length; aidx += 1) {
-      result[apps[aidx]] = (
-        await matsCollections.variable.findOneAsync({
-          name: "variable",
-        })
-      ).optionsMap[apps[aidx]];
-      if (
-        typeof result[apps[aidx]] !== "string" &&
-        !(result[apps[aidx]] instanceof String)
-      )
-        result[apps[aidx]] = result[apps[aidx]].sumsDB;
-    }
-  } else {
-    result[(await matsCollections.Settings.findOneAsync()).Title] = (
-      await matsCollections.Databases.findOneAsync({
-        role: matsTypes.DatabaseRoles.SUMS_DATA,
-        status: "active",
-      })
-    ).database;
-  }
-  return result;
-}
-
-// helper function for getting a metadata map from a MATS selector, keyed by app title and model display text
-async function getMapByAppAndModel(selector, mapType) {
-  let flatJSON = "";
-  try {
-    let result;
-    if (
-      matsCollections[selector] !== undefined &&
-      (await matsCollections[selector].findOneAsync({ name: selector })) !==
-        undefined &&
-      (await matsCollections[selector].findOneAsync({ name: selector }))[mapType] !==
-        undefined
-    ) {
-      // get map of requested selector's metadata
-      result = (
-        await matsCollections[selector].findOneAsync({
-          name: selector,
-        })
-      )[mapType];
-      let newResult = {};
-      if (
-        mapType === "valuesMap" ||
-        selector === "variable" ||
-        selector === "statistic"
-      ) {
-        // valueMaps always need to be re-keyed by app (statistic and variable get their valuesMaps from optionsMaps)
-        newResult = await mapMapToApps(result);
-        result = newResult;
-      } else if (
-        matsCollections.database === undefined &&
-        !(
-          matsCollections.variable !== undefined &&
-          matsCollections.threshold !== undefined
-        )
-      ) {
-        // key by app title if we're not already
-        const appTitle = (await matsCollections.Settings.findOneAsync()).Title;
-        newResult[appTitle] = result;
-        result = newResult;
-      }
-    } else {
-      result = {};
-    }
-    flatJSON = JSON.stringify(result);
-  } catch (e) {
-    console.log(`error retrieving metadata from ${selector}: `, e);
-    flatJSON = JSON.stringify({
-      error: e,
-    });
-  }
-  return flatJSON;
-}
-
-// helper function for getting a date metadata map from a MATS selector, keyed by app title and model display text
-async function getDateMapByAppAndModel() {
-  let flatJSON = "";
-  try {
-    let result;
-    // the date map can be in a few places. we have to hunt for it.
-    if (
-      matsCollections.database !== undefined &&
-      (await matsCollections.database.findOneAsync({
-        name: "database",
-      })) !== undefined &&
-      (
-        await matsCollections.database.findOneAsync({
-          name: "database",
-        })
-      ).dates !== undefined
-    ) {
-      result = (
-        await matsCollections.database.findOneAsync({
-          name: "database",
-        })
-      ).dates;
-    } else if (
-      matsCollections.variable !== undefined &&
-      (await matsCollections.variable.findOneAsync({
-        name: "variable",
-      })) !== undefined &&
-      (
-        await matsCollections.variable.findOneAsync({
-          name: "variable",
-        })
-      ).dates !== undefined
-    ) {
-      result = (
-        await matsCollections.variable.findOneAsync({
-          name: "variable",
-        })
-      ).dates;
-    } else if (
-      matsCollections["data-source"] !== undefined &&
-      (await matsCollections["data-source"].findOneAsync({
-        name: "data-source",
-      })) !== undefined &&
-      (
-        await matsCollections["data-source"].findOneAsync({
-          name: "data-source",
-        })
-      ).dates !== undefined
-    ) {
-      result = (
-        await matsCollections["data-source"].findOneAsync({
-          name: "data-source",
-        })
-      ).dates;
-    } else {
-      result = {};
-    }
-    if (
-      matsCollections.database === undefined &&
-      !(
-        matsCollections.variable !== undefined &&
-        matsCollections.threshold !== undefined
-      )
-    ) {
-      // key by app title if we're not already
-      const appTitle = (await matsCollections.Settings.findOneAsync()).Title;
-      const newResult = {};
-      newResult[appTitle] = result;
-      result = newResult;
-    }
-    flatJSON = JSON.stringify(result);
-  } catch (e) {
-    console.log("error retrieving datemap", e);
-    flatJSON = JSON.stringify({
-      error: e,
-    });
-  }
-  return flatJSON;
-}
-
-// helper function for getting a metadata map from a MATS selector, keyed by app title
-async function getMapByApp(selector) {
-  let flatJSON = "";
-  try {
-    let result;
-    if (
-      matsCollections[selector] !== undefined &&
-      (await matsCollections[selector].findOneAsync({ name: selector })) !== undefined
-    ) {
-      // get array of requested selector's metadata
-      result = (
-        await matsCollections[selector].findOneAsync({
-          name: selector,
-        })
-      ).options;
-      if (!Array.isArray(result)) result = Object.keys(result);
-    } else if (selector === "statistic") {
-      result = ["ACC"];
-    } else if (selector === "variable") {
-      result = [(await matsCollections.Settings.findOneAsync()).Title];
-    } else {
-      result = [];
-    }
-    // put results in a map keyed by app
-    let newResult;
-    if (result.length === 0) {
-      newResult = {};
-    } else {
-      newResult = await mapArrayToApps(result);
-    }
-    flatJSON = JSON.stringify(newResult);
-  } catch (e) {
-    console.log(`error retrieving metadata from ${selector}: `, e);
-    flatJSON = JSON.stringify({
-      error: e,
-    });
-  }
-  return flatJSON;
-}
-
-// helper function for populating the levels in a MATS selector
-async function getlevelsByApp() {
-  let flatJSON = "";
-  try {
-    let result;
-    if (
-      matsCollections.level !== undefined &&
-      (await matsCollections.level.findOneAsync({ name: "level" })) !== undefined
-    ) {
-      // we have levels already defined
-      result = (
-        await matsCollections.level.findOneAsync({
-          name: "level",
-        })
-      ).options;
-      if (!Array.isArray(result)) result = Object.keys(result);
-    } else if (
-      matsCollections.top !== undefined &&
-      (await matsCollections.top.findOneAsync({ name: "top" })) !== undefined
-    ) {
-      // use the MATS mandatory levels
-      result = _.range(100, 1050, 50);
-      if (!Array.isArray(result)) result = Object.keys(result);
-    } else {
-      result = [];
-    }
-    let newResult;
-    if (result.length === 0) {
-      newResult = {};
-    } else {
-      newResult = await mapArrayToApps(result);
-    }
-    flatJSON = JSON.stringify(newResult);
-  } catch (e) {
-    console.log("error retrieving levels: ", e);
-    flatJSON = JSON.stringify({
-      error: e,
-    });
-  }
-  return flatJSON;
-}
-
-// private middleware for getApps route
-const getApps = async function (res) {
-  // this function returns an array of apps.
-  if (Meteor.isServer) {
-    let flatJSON = "";
-    try {
-      const result = await getListOfApps();
-      flatJSON = JSON.stringify(result);
-    } catch (e) {
-      console.log("error retrieving apps: ", e);
-      flatJSON = JSON.stringify({
-        error: e,
-      });
-    }
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getAppSumsDBs route
-const getAppSumsDBs = async function (res) {
-  // this function returns map of apps and appRefs.
-  if (Meteor.isServer) {
-    let flatJSON = "";
-    try {
-      const result = await getListOfAppDBs();
-      flatJSON = JSON.stringify(result);
-    } catch (e) {
-      console.log("error retrieving apps: ", e);
-      flatJSON = JSON.stringify({
-        error: e,
-      });
-    }
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getModels route
-const getModels = async function (res) {
-  // this function returns a map of models keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("data-source", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getRegions route
-const getRegions = async function (res) {
-  // this function returns a map of regions keyed by app title and model display text
-  if (Meteor.isServer) {
-    let flatJSON = await getMapByAppAndModel("region", "optionsMap");
-    if (flatJSON === "{}") {
-      flatJSON = await getMapByAppAndModel("vgtyp", "optionsMap");
-    }
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getRegionsValuesMap route
-const getRegionsValuesMap = async function (res) {
-  // this function returns a map of regions values keyed by app title
-  if (Meteor.isServer) {
-    let flatJSON = await getMapByAppAndModel("region", "valuesMap");
-    if (flatJSON === "{}") {
-      flatJSON = await getMapByAppAndModel("vgtyp", "valuesMap");
-    }
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getStatistics route
-const getStatistics = async function (res) {
-  // this function returns an map of statistics keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByApp("statistic");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getStatisticsValuesMap route
-const getStatisticsValuesMap = async function (res) {
-  // this function returns a map of statistic values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("statistic", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getVariables route
-const getVariables = async function (res) {
-  // this function returns an map of variables keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByApp("variable");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getVariablesValuesMap route
-const getVariablesValuesMap = async function (res) {
-  // this function returns a map of variable values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("variable", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getThresholds route
-const getThresholds = async function (res) {
-  // this function returns a map of thresholds keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("threshold", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getThresholdsValuesMap route
-const getThresholdsValuesMap = async function (res) {
-  // this function returns a map of threshold values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("threshold", "valuesMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getScales route
-const getScales = async function (res) {
-  // this function returns a map of scales keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("scale", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getScalesValuesMap route
-const getScalesValuesMap = async function (res) {
-  // this function returns a map of scale values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("scale", "valuesMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getTruth route
-const getTruths = async function (res) {
-  // this function returns a map of truths keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("truth", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getTruthValuesMap route
-const getTruthsValuesMap = async function (res) {
-  // this function returns a map of truth values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("truth", "valuesMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getFcstLengths route
-const getFcstLengths = async function (res) {
-  // this function returns a map of forecast lengths keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("forecast-length", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getFcstTypes route
-const getFcstTypes = async function (res) {
-  // this function returns a map of forecast types keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("forecast-type", "optionsMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getFcstTypesValuesMap route
-const getFcstTypesValuesMap = async function (res) {
-  // this function returns a map of forecast type values keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByAppAndModel("forecast-type", "valuesMap");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getValidTimes route
-const getValidTimes = async function (res) {
-  // this function returns an map of valid times keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getMapByApp("valid-time");
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getValidTimes route
-const getLevels = async function (res) {
-  // this function returns an map of pressure levels keyed by app title
-  if (Meteor.isServer) {
-    const flatJSON = await getlevelsByApp();
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private middleware for getDates route
-const getDates = async function (res) {
-  // this function returns a map of dates keyed by app title and model display text
-  if (Meteor.isServer) {
-    const flatJSON = await getDateMapByAppAndModel();
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// helper function for getCSV
-const stringifyCurveData = function (stringify, dataArray, res) {
-  const thisDataArray = dataArray[0];
-  stringify.stringify(
-    thisDataArray,
-    {
-      header: true,
-    },
-    function (err, output) {
-      if (err) {
-        console.log("error in getCSV:", err);
-        res.write(`error,${err.toLocaleString()}`);
-        res.end(`<body><h1>getCSV Error! ${err.toLocaleString()}</h1></body>`);
-        return;
-      }
-      res.write(output);
-      if (dataArray.length > 1) {
-        const newDataArray = dataArray.slice(1);
-        stringifyCurveData(stringify, newDataArray, res);
-      } else {
-        res.end();
-      }
-    }
-  );
-};
-
-// private method for getting pagenated data
-// a newPageIndex of -1000 means get all the data (used for export)
-// a newPageIndex of -2000 means get just the last page
-const getPagenatedData = function (rky, p, np) {
-  if (Meteor.isServer) {
-    const key = rky;
-    const myPageIndex = p;
-    const newPageIndex = np;
-    let rawReturn;
-
-    try {
-      const result = matsCache.getResult(key);
-      rawReturn = result === undefined ? undefined : result.result; // getResult structure is {key:something, result:resultObject}
-    } catch (e) {
-      console.log("getPagenatedData: Error - ", e);
-      return undefined;
-    }
-    const ret =
-      rawReturn === undefined ? undefined : JSON.parse(JSON.stringify(rawReturn));
-    let start;
-    let end;
-    let direction = 1;
-    if (newPageIndex === -1000) {
-      // all the data
-      start = 0;
-      end = Number.MAX_VALUE;
-    } else if (newPageIndex === -2000) {
-      // just the last page
-      start = -2000;
-      direction = -1;
-    } else if (myPageIndex <= newPageIndex) {
-      // proceed forward
-      start = (newPageIndex - 1) * 100;
-      end = newPageIndex * 100;
-    } else {
-      // move back
-      direction = -1;
-      start = newPageIndex * 100;
-      end = (newPageIndex + 1) * 100;
-    }
-
-    let dsiStart;
-    let dsiEnd;
-    for (let csi = 0; csi < ret.data.length; csi += 1) {
-      if (ret.data[csi].x && ret.data[csi].x.length > 100) {
-        dsiStart = start;
-        dsiEnd = end;
-        if (dsiStart > ret.data[csi].x.length || dsiStart === -2000) {
-          // show the last page if we either requested it specifically or are trying to navigate past it
-          dsiStart = Math.floor(rawReturn.data[csi].x.length / 100) * 100;
-          dsiEnd = rawReturn.data[csi].x.length;
-          if (dsiEnd === dsiStart) {
-            // make sure the last page isn't empty -= 1if rawReturn.data[csi].data.length/100 produces a whole number,
-            // dsiStart and dsiEnd would be the same. This makes sure that the last full page is indeed the last page, without a phantom empty page afterwards
-            dsiStart = dsiEnd - 100;
-          }
-        }
-        if (dsiStart < 0) {
-          // show the first page if we are trying to navigate before it
-          dsiStart = 0;
-          dsiEnd = 100;
-        }
-        if (dsiEnd < dsiStart) {
-          // make sure that the end is after the start
-          dsiEnd = dsiStart + 100;
-        }
-        if (dsiEnd > ret.data[csi].x.length) {
-          // make sure we don't request past the end  -= 1 if results are one page, this should convert the
-          // start and end from 0 and 100 to 0 and whatever the end is.
-          dsiEnd = ret.data[csi].x.length;
-        }
-        ret.data[csi].x = rawReturn.data[csi].x.slice(dsiStart, dsiEnd);
-        ret.data[csi].y = rawReturn.data[csi].y.slice(dsiStart, dsiEnd);
-        ret.data[csi].stats = rawReturn.data[csi].stats.slice(dsiStart, dsiEnd);
-        ret.data[csi].glob_stats = rawReturn.data[csi].glob_stats;
-      }
-    }
-
-    if (direction === 1) {
-      ret.dsiRealPageIndex = Math.floor(dsiEnd / 100);
-    } else {
-      ret.dsiRealPageIndex = Math.floor(dsiStart / 100);
-    }
-    ret.dsiTextDirection = direction;
-    return ret;
-  }
-  return null;
-};
-
-// private method for getting pagenated results and flattening them in order to be appropriate for text display.
-const getFlattenedResultData = function (rk, p, np) {
-  if (Meteor.isServer) {
-    try {
-      const r = rk;
-      const thisP = p;
-      const thisNP = np;
-      // get the pagenated data
-      const result = getPagenatedData(r, thisP, thisNP);
-      // find the type
-      const { plotTypes } = result.basis.plotParams;
-      const plotType = _.invert(plotTypes).true;
-      // extract data
-      let isCTC = false;
-      let isModeSingle = false;
-      let isModePairs = false;
-      let labelSuffix;
-      const returnData = {};
-      const stats = {};
-      let curveData = []; // array of maps
-      const { data } = result;
-      const { dsiRealPageIndex } = result;
-      const { dsiTextDirection } = result;
-      let firstBestFitIndex = -1;
-      let bestFitIndexes = {};
-      switch (plotType) {
-        case matsTypes.PlotTypes.timeSeries:
-        case matsTypes.PlotTypes.profile:
-        case matsTypes.PlotTypes.dieoff:
-        case matsTypes.PlotTypes.threshold:
-        case matsTypes.PlotTypes.validtime:
-        case matsTypes.PlotTypes.dailyModelCycle:
-        case matsTypes.PlotTypes.gridscale:
-        case matsTypes.PlotTypes.yearToYear:
-          switch (plotType) {
-            case matsTypes.PlotTypes.timeSeries:
-            case matsTypes.PlotTypes.dailyModelCycle:
-              labelSuffix = " time";
-              break;
-            case matsTypes.PlotTypes.profile:
-              labelSuffix = " level";
-              break;
-            case matsTypes.PlotTypes.dieoff:
-              labelSuffix = " forecast lead time";
-              break;
-            case matsTypes.PlotTypes.validtime:
-              labelSuffix = " hour of day";
-              break;
-            case matsTypes.PlotTypes.threshold:
-              labelSuffix = " threshold";
-              break;
-            case matsTypes.PlotTypes.gridscale:
-              labelSuffix = " grid scale";
-              break;
-            case matsTypes.PlotTypes.yearToYear:
-              labelSuffix = " year";
-              break;
-            default:
-              labelSuffix = "x-value";
-          }
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (reservedWords.indexOf(data[ci].label) === -1) {
-              // for each curve
-              isCTC =
-                data[ci] !== undefined &&
-                ((data[ci].stats !== undefined &&
-                  data[ci].stats[0] !== undefined &&
-                  data[ci].stats[0].hit !== undefined) ||
-                  (data[ci].hitTextOutput !== undefined &&
-                    data[ci].hitTextOutput.length > 0));
-              isModePairs =
-                data[ci] !== undefined &&
-                data[ci].stats !== undefined &&
-                data[ci].stats[0] !== undefined &&
-                data[ci].stats[0].avgInterest !== undefined;
-              isModeSingle =
-                data[ci] !== undefined &&
-                data[ci].stats !== undefined &&
-                data[ci].stats[0] !== undefined &&
-                data[ci].stats[0].nForecast !== undefined;
-              // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-              stats.label = data[ci].label;
-              stats.mean = data[ci].glob_stats.dMean;
-              stats["standard deviation"] = data[ci].glob_stats.sd;
-              stats.n = data[ci].glob_stats.nGood;
-              if (
-                plotType === matsTypes.PlotTypes.timeSeries ||
-                plotType === matsTypes.PlotTypes.profile
-              ) {
-                stats["standard error"] = data[ci].glob_stats.stdeBetsy;
-                stats.lag1 = data[ci].glob_stats.lag1;
-              }
-              stats.minimum = data[ci].glob_stats.minVal;
-              stats.maximum = data[ci].glob_stats.maxVal;
-              returnData.stats[data[ci].label] = JSON.parse(JSON.stringify(stats));
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].x.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                if (plotType === matsTypes.PlotTypes.profile) {
-                  curveDataElement[data[ci].label + labelSuffix] = data[ci].y[cdi];
-                } else {
-                  curveDataElement[data[ci].label + labelSuffix] = data[ci].x[cdi];
-                }
-                if (isCTC) {
-                  curveDataElement.stat = data[ci].stats[cdi].stat;
-                  curveDataElement.n = data[ci].stats[cdi].n;
-                  curveDataElement.hit = data[ci].stats[cdi].hit;
-                  curveDataElement.fa = data[ci].stats[cdi].fa;
-                  curveDataElement.miss = data[ci].stats[cdi].miss;
-                  curveDataElement.cn = data[ci].stats[cdi].cn;
-                } else if (isModeSingle) {
-                  curveDataElement.stat = data[ci].stats[cdi].stat;
-                  curveDataElement.nForecast = data[ci].stats[cdi].nForecast;
-                  curveDataElement.nMatched = data[ci].stats[cdi].nMatched;
-                  curveDataElement.nSimple = data[ci].stats[cdi].nSimple;
-                  curveDataElement.nTotal = data[ci].stats[cdi].nTotal;
-                } else if (isModePairs) {
-                  curveDataElement.stat = data[ci].stats[cdi].stat;
-                  curveDataElement.n = data[ci].stats[cdi].n;
-                  curveDataElement.avgInterest = data[ci].stats[cdi].avgInterest;
-                } else {
-                  curveDataElement.stat = data[ci].stats[cdi].stat;
-                  curveDataElement.mean = data[ci].stats[cdi].mean;
-                  curveDataElement["std dev"] = data[ci].stats[cdi].sd;
-                  if (
-                    plotType === matsTypes.PlotTypes.timeSeries ||
-                    plotType === matsTypes.PlotTypes.profile
-                  ) {
-                    curveDataElement["std error"] = data[ci].stats[cdi].stdeBetsy;
-                    curveDataElement.lag1 = data[ci].stats[cdi].lag1;
-                  }
-                  curveDataElement.n = data[ci].stats[cdi].nGood;
-                }
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.reliability:
-        case matsTypes.PlotTypes.roc:
-        case matsTypes.PlotTypes.performanceDiagram:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            // for each curve
-            // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (
-              reservedWords.indexOf(data[ci].label) === -1 &&
-              !data[ci].label.includes(matsTypes.ReservedWords.noSkill)
-            ) {
-              stats.label = data[ci].label;
-              if (plotType === matsTypes.PlotTypes.reliability) {
-                stats["sample climo"] = data[ci].glob_stats.sample_climo;
-              } else if (plotType === matsTypes.PlotTypes.roc) {
-                stats.auc = data[ci].glob_stats.auc;
-              }
-              returnData.stats[data[ci].label] = JSON.parse(JSON.stringify(stats));
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].y.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                if (plotType === matsTypes.PlotTypes.reliability) {
-                  curveDataElement[`${data[ci].label} probability bin`] =
-                    data[ci].stats[cdi].prob_bin;
-                  if (data[ci].stats[cdi].hit_rate) {
-                    curveDataElement["hit rate"] = data[ci].stats[cdi].hit_rate;
-                  } else {
-                    curveDataElement["observed frequency"] =
-                      data[ci].stats[cdi].obs_freq;
-                  }
-                } else {
-                  curveDataElement[`${data[ci].label} bin value`] =
-                    data[ci].stats[cdi].bin_value;
-                  curveDataElement["probability of detection"] =
-                    data[ci].stats[cdi].pody;
-                  if (plotType === matsTypes.PlotTypes.roc) {
-                    curveDataElement["probability of false detection"] =
-                      data[ci].stats[cdi].pofd;
-                  } else {
-                    curveDataElement["success ratio"] = data[ci].stats[cdi].fa;
-                  }
-                  curveDataElement.n = data[ci].stats[cdi].n;
-                }
-                if (data[ci].stats[cdi].obs_y) {
-                  curveDataElement.oy = data[ci].stats[cdi].obs_y;
-                  curveDataElement.on = data[ci].stats[cdi].obs_n;
-                } else {
-                  curveDataElement.hitcount = data[ci].stats[cdi].hit_count;
-                  curveDataElement.fcstcount = data[ci].stats[cdi].fcst_count;
-                }
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.gridscaleProb:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            // for each curve
-            // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (reservedWords.indexOf(data[ci].label) === -1) {
-              stats.label = data[ci].label;
-              returnData.stats[data[ci].label] = JSON.parse(JSON.stringify(stats));
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].y.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                curveDataElement[`${data[ci].label} probability bin`] =
-                  data[ci].stats[cdi].bin_value;
-                curveDataElement["number of grid points"] = data[ci].stats[cdi].n_grid;
-                curveDataElement.n = data[ci].stats[cdi].n;
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.simpleScatter:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            // for each curve
-            // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (reservedWords.indexOf(data[ci].label) === -1) {
-              stats.label = data[ci].label;
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].y.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                curveDataElement[`${data[ci].label} bin value`] =
-                  data[ci].stats[cdi].bin_value;
-                curveDataElement["x-stat"] = data[ci].stats[cdi].xstat;
-                curveDataElement["y-stat"] = data[ci].stats[cdi].ystat;
-                curveDataElement.n = data[ci].stats[cdi].n;
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.map:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          stats.label = data[0].label;
-          stats["total number of obs"] = data[0].stats.reduce(function (prev, curr) {
-            return prev + curr.nTimes;
-          }, 0);
-          stats["mean difference"] = matsDataUtils.average(data[0].queryVal);
-          stats["standard deviation"] = matsDataUtils.stdev(data[0].queryVal);
-          stats["minimum time"] = data[0].stats.reduce(function (prev, curr) {
-            return prev < curr.min_time ? prev : curr.min_time;
-          });
-          stats["minimum time"] = moment
-            .utc(stats["minimum time"] * 1000)
-            .format("YYYY-MM-DD HH:mm");
-          stats["maximum time"] = data[0].stats.reduce(function (prev, curr) {
-            return prev > curr.max_time ? prev : curr.max_time;
-          });
-          stats["maximum time"] = moment
-            .utc(stats["maximum time"] * 1000)
-            .format("YYYY-MM-DD HH:mm");
-
-          returnData.stats[data[0].label] = JSON.parse(JSON.stringify(stats));
-
-          isCTC =
-            data[0] !== undefined &&
-            data[0].stats !== undefined &&
-            data[0].stats[0] !== undefined &&
-            data[0].stats[0].hit !== undefined;
-
-          curveData = []; // array of maps
-          for (let si = 0; si < data[0].siteName.length; si += 1) {
-            const curveDataElement = {};
-            curveDataElement["site name"] = data[0].siteName[si];
-            curveDataElement["number of times"] = data[0].stats[si].nTimes;
-            if (isCTC) {
-              curveDataElement.stat = data[0].queryVal[si];
-              curveDataElement.hit = data[0].stats[si].hit;
-              curveDataElement.fa = data[0].stats[si].fa;
-              curveDataElement.miss = data[0].stats[si].miss;
-              curveDataElement.cn = data[0].stats[si].cn;
-            } else {
-              curveDataElement["start date"] = moment
-                .utc(data[0].stats[si].min_time * 1000)
-                .format("YYYY-MM-DD HH:mm");
-              curveDataElement["end date"] = moment
-                .utc(data[0].stats[si].max_time * 1000)
-                .format("YYYY-MM-DD HH:mm");
-              curveDataElement.stat = data[0].queryVal[si];
-            }
-            curveData.push(curveDataElement);
-          }
-          returnData.data[data[0].label] = curveData;
-          break;
-        case matsTypes.PlotTypes.histogram:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            // for each curve
-            // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (reservedWords.indexOf(data[ci].label) === -1) {
-              stats.label = data[ci].label;
-              stats.mean = data[ci].glob_stats.glob_mean;
-              stats["standard deviation"] = data[ci].glob_stats.glob_sd;
-              stats.n = data[ci].glob_stats.glob_n;
-              stats.minimum = data[ci].glob_stats.glob_min;
-              stats.maximum = data[ci].glob_stats.glob_max;
-              returnData.stats[data[ci].label] = JSON.parse(JSON.stringify(stats));
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].x.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                curveDataElement[`${data[ci].label} bin range`] =
-                  data[ci].bin_stats[cdi].binLabel;
-                curveDataElement.n = data[ci].bin_stats[cdi].bin_n;
-                curveDataElement["bin rel freq"] = data[ci].bin_stats[cdi].bin_rf;
-                curveDataElement["bin lower bound"] =
-                  data[ci].bin_stats[cdi].binLowBound;
-                curveDataElement["bin upper bound"] =
-                  data[ci].bin_stats[cdi].binUpBound;
-                curveDataElement["bin mean"] = data[ci].bin_stats[cdi].bin_mean;
-                curveDataElement["bin std dev"] = data[ci].bin_stats[cdi].bin_sd;
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.ensembleHistogram:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            // for each curve
-            // if the curve label is a reserved word do not process the curve (its a zero or max curve)
-            const reservedWords = Object.values(matsTypes.ReservedWords);
-            if (reservedWords.indexOf(data[ci].label) === -1) {
-              stats.label = data[ci].label;
-              stats.mean = data[ci].glob_stats.dMean;
-              stats["standard deviation"] = data[ci].glob_stats.sd;
-              stats.n = data[ci].glob_stats.nGood;
-              stats.minimum = data[ci].glob_stats.minVal;
-              stats.maximum = data[ci].glob_stats.maxVal;
-              returnData.stats[data[ci].label] = JSON.parse(JSON.stringify(stats));
-
-              curveData = []; // array of maps
-              for (let cdi = 0; cdi < data[ci].x.length; cdi += 1) {
-                // for each datapoint
-                const curveDataElement = {};
-                curveDataElement[`${data[ci].label} bin`] = data[ci].x[cdi];
-                curveDataElement.n = data[ci].bin_stats[cdi].bin_n;
-                curveDataElement["bin rel freq"] = data[ci].bin_stats[cdi].bin_rf;
-                curveData.push(curveDataElement);
-              }
-              returnData.data[data[ci].label] = curveData;
-            }
-          }
-          break;
-        case matsTypes.PlotTypes.contour:
-        case matsTypes.PlotTypes.contourDiff:
-          returnData.stats = {}; // map of maps
-          returnData.data = {}; // map of arrays of maps
-          stats.label = data[0].label;
-          stats["total number of points"] = data[0].glob_stats.n;
-          stats["mean stat"] = data[0].glob_stats.mean;
-          stats["minimum time"] = moment
-            .utc(data[0].glob_stats.minDate * 1000)
-            .format("YYYY-MM-DD HH:mm");
-          stats["maximum time"] = moment
-            .utc(data[0].glob_stats.maxDate * 1000)
-            .format("YYYY-MM-DD HH:mm");
-
-          returnData.stats[data[0].label] = JSON.parse(JSON.stringify(stats));
-
-          isCTC =
-            data[0] !== undefined &&
-            data[0].hitTextOutput !== undefined &&
-            data[0].hitTextOutput.length > 0;
-
-          curveData = []; // array of maps
-          for (let si = 0; si < data[0].xTextOutput.length; si += 1) {
-            const curveDataElement = {};
-            curveDataElement.xVal = data[0].xTextOutput[si];
-            curveDataElement.yVal = data[0].yTextOutput[si];
-            curveDataElement.stat = data[0].zTextOutput[si];
-            curveDataElement.N = data[0].nTextOutput[si];
-            if (isCTC) {
-              curveDataElement.hit = data[0].hitTextOutput[si];
-              curveDataElement.fa = data[0].faTextOutput[si];
-              curveDataElement.miss = data[0].missTextOutput[si];
-              curveDataElement.cn = data[0].cnTextOutput[si];
-            } else {
-              curveDataElement["Start Date"] = moment
-                .utc(data[0].minDateTextOutput[si] * 1000)
-                .format("YYYY-MM-DD HH:mm");
-              curveDataElement["End Date"] = moment
-                .utc(data[0].maxDateTextOutput[si] * 1000)
-                .format("YYYY-MM-DD HH:mm");
-            }
-            curveData.push(curveDataElement);
-          }
-          returnData.data[data[0].label] = curveData;
-          break;
-        case matsTypes.PlotTypes.scatter2d:
-          firstBestFitIndex = -1;
-          bestFitIndexes = {};
-
-          curveData = []; // array of maps
-          for (let ci = 0; ci < data.length; ci += 1) {
-            if (ci === firstBestFitIndex) {
-              break; // best fit curves are at the end so do not do further processing
-            }
-            curveData = data[ci];
-            // look for a best fit curve - only have to look at curves with higher index than this one
-            for (let cbi = ci + 1; cbi < data.length; cbi += 1) {
-              if (
-                data[cbi].label.indexOf(curveData.label) !== -1 &&
-                data[cbi].label.indexOf("-best fit") !== -1
-              ) {
-                bestFitIndexes[ci] = cbi;
-                if (firstBestFitIndex === -1) {
-                  firstBestFitIndex = cbi;
-                }
-                break;
-              }
-            }
-            const curveTextData = [];
-            for (let cdi = 0; cdi < curveData.data.length; cdi += 1) {
-              const element = {};
-              [element.xAxis] = curveData.data[cdi];
-              [, element.yAxis] = curveData.data[cdi];
-              if (bestFitIndexes[ci] === undefined) {
-                element["best fit"] = "none;";
-              } else {
-                [, element["best fit"]] = data[bestFitIndexes[ci]].data[cdi];
-              }
-              curveTextData.push(element);
-            }
-            returnData[curveData.label] = curveTextData;
-          }
-          break;
-        default:
-          return undefined;
-      }
-      returnData.dsiRealPageIndex = dsiRealPageIndex;
-      returnData.dsiTextDirection = dsiTextDirection;
-      return returnData;
-    } catch (error) {
-      throw new Meteor.Error(
-        `Error in getFlattenedResultData function: ${error.message}`
-      );
-    }
-  }
-  return null;
-};
-
-// private middleware for getCSV route
-const getCSV = function (params, res) {
-  if (Meteor.isServer) {
-    const stringify = require("csv-stringify");
-    let csv = "";
-    try {
-      const result = getFlattenedResultData(params.key, 0, -1000);
-      const statArray = Object.values(result.stats);
-      const dataArray = Object.values(result.data);
-
-      const fileName = `matsplot-${moment.utc().format("YYYYMMDD-HH.mm.ss")}.csv`;
-      res.setHeader("Content-disposition", `attachment; filename=${fileName}`);
-      res.setHeader("Content-Type", "attachment.ContentType");
-      stringify.stringify(
-        statArray,
-        {
-          header: true,
-        },
-        function (err, output) {
-          if (err) {
-            console.log("error in getCSV:", err);
-            res.write(`error,${err.toLocaleString()}`);
-            res.end(`<body><h1>getCSV Error! ${err.toLocaleString()}</h1></body>`);
-            return;
-          }
-          res.write(output);
-          stringifyCurveData(stringify, dataArray, res);
-        }
-      );
-    } catch (e) {
-      console.log("error retrieving data: ", e);
-      csv = `error,${e.toLocaleString()}`;
-      res.setHeader("Content-disposition", "attachment; filename=matsplot.csv");
-      res.setHeader("Content-Type", "attachment.ContentType");
-      res.end(`<body><h1>getCSV Error! ${csv}</h1></body>`);
-    }
-  }
-};
-
-// private middleware for getJSON route
-const getJSON = function (params, res) {
-  if (Meteor.isServer) {
-    let flatJSON = "";
-    try {
-      const result = getPagenatedData(params.key, 0, -1000);
-      flatJSON = JSON.stringify(result.basis);
-    } catch (e) {
-      console.log("error retrieving data: ", e);
-      flatJSON = JSON.stringify({
-        error: e,
-      });
-    }
-    res.setHeader("Content-Type", "application/json");
-    res.write(flatJSON);
-    res.end();
-  }
-};
-
-// private define a middleware for refreshing the metadata
-const refreshMetadataMWltData = async function (res) {
-  if (Meteor.isServer) {
-    console.log("Server route asked to refresh metadata");
-
-    try {
-      console.log("GUI asked to refresh metadata");
-      await checkMetaDataRefresh();
-    } catch (e) {
-      console.log(e);
-      res.end(
-        `<body>` +
-          `<h1>refreshMetadata Failed!</h1>` +
-          `<p>${e.message}</p>` +
-          `</body>`
-      );
-    }
-    res.end("<body><h1>refreshMetadata Done!</h1></body>");
-  }
-};
-
-// private middleware for causing the scorecard to refresh its mongo collection for a given document
-const refreshScorecard = function (params, res) {
-  if (Meteor.isServer) {
-    const docId = decodeURIComponent(params.docId);
-    // get userName, name, submitted, processedAt from id
-    // SC:anonymous -= 1submitted:20230322230435 -= 11block:0:02/19/2023_20_00_-_03/21/2023_20_00
-    if (global.cbScorecardPool === undefined) {
-      throw new Meteor.Error("getScorecardData: No cbScorecardPool defined");
-    }
-    const statement = `SELECT sc.*
-                From
-                    vxdata._default.SCORECARD sc
-                WHERE
-                    sc.id='${docId}';`;
-    global.cbScorecardPool
-      .queryCB(statement)
-      .then(async (result) => {
-        // insert this result into the mongo Scorecard collection - createdAt is used for TTL
-        // created at gets updated each display even if it already existed.
-        // TTL is 24 hours
-        if (typeof result === "string") {
-          throw new Error(`Error from couchbase query - ${result}`);
-        } else if (result[0] === undefined) {
-          throw new Error("Error from couchbase query - document not found");
-        } else {
-          await matsCollections.Scorecard.upsertAsync(
-            {
-              "scorecard.userName": result[0].userName,
-              "scorecard.name": result[0].name,
-              "scorecard.submitted": result[0].submitted,
-              "scorecard.processedAt": result[0].processedAt,
-            },
-            {
-              $set: {
-                createdAt: new Date(),
-                scorecard: result[0],
-              },
-            }
-          );
-        }
-        res.end("<body><h1>refreshScorecard Done!</h1></body>");
-      })
-      .catch((err) => {
-        res.end(
-          `<body>` +
-            `<h1>refreshScorecard Failed!</h1>` +
-            `<p>${err.message}</p>` +
-            `</body>`
-        );
-      });
-  }
-};
-
-const setStatusScorecard = function (params, req, res) {
-  if (Meteor.isServer) {
-    const docId = decodeURIComponent(params.docId);
-    let body = "";
-    req.on(
-      "data",
-      Meteor.bindEnvironment(function (data) {
-        body += data;
-      })
-    );
-
-    req.on(
-      "end",
-      Meteor.bindEnvironment(async function () {
-        // console.log(body);
-        try {
-          const doc = JSON.parse(body);
-          const docStatus = doc.status;
-          const found = await matsCollections.Scorecard.find({
-            id: docId,
-          }).fetchAsync();
-          if (found.length === 0) {
-            throw new Error("Error from scorecard lookup - document not found");
-          }
-          await matsCollections.Scorecard.upsertAsync(
-            {
-              id: docId,
-            },
-            {
-              $set: {
-                docStatus,
-              },
-            }
-          );
-          // set error if there is one somehow. (use the session?)
-          res.end("<body><h1>setScorecardStatus Done!</h1></body>");
-        } catch (err) {
-          res.statusCode = 400;
-          res.end(
-            `<body>` +
-              `<h1>setScorecardStatus Failed!</h1>` +
-              `<p>${err.message}</p>` +
-              `</body>`
-          );
-        }
-      })
-    );
   }
 };
 
@@ -2949,247 +1460,8 @@ if (Meteor.isServer) {
     }
   ); // 15 min expiration
 
-  // set the default proxy prefix path to ""
-  // If the settings are not complete, they will be set by the configuration and written out, which will cause the app to reset
-  if (Meteor.settings.public && !Meteor.settings.public.proxy_prefix_path) {
-    Meteor.settings.public.proxy_prefix_path = "";
-  }
-
-  const proxyPrefixPath = Meteor.settings.public.proxy_prefix_path;
-
-  // needed for middleware routes
-  const app = WebApp.express();
-  const router = WebApp.express.Router();
-
-  router.get("/healthz", async (req, res) => {
-    await getHealth(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/healthz`, async (req, res) => {
-    await getHealth(res);
-  });
-
-  router.get("/CSV/:f/:key/:m/:a", (req, res) => {
-    getCSV(req.params, res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/CSV/:f/:key/:m/:a`, (req, res) => {
-    getCSV(req.params, res);
-  });
-
-  router.get("/JSON/:f/:key/:m/:a", (req, res) => {
-    getJSON(req.params, res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/JSON/:f/:key/:m/:a`, (req, res) => {
-    getJSON(req.params, res);
-  });
-
-  router.get("/clearCache", (req, res) => {
-    clearCache(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/clearCache`, (req, res) => {
-    clearCache(res);
-  });
-
-  router.get("/getApps", async (req, res) => {
-    await getApps(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getApps`, async (req, res) => {
-    await getApps(res);
-  });
-
-  router.get("/getAppSumsDBs", async (req, res) => {
-    await getAppSumsDBs(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getAppSumsDBs`, async (req, res) => {
-    await getAppSumsDBs(res);
-  });
-
-  router.get("/getModels", async (req, res) => {
-    await getModels(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getModels`, async (req, res) => {
-    await getModels(res);
-  });
-
-  router.get("/getRegions", async (req, res) => {
-    await getRegions(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getRegions`, async (req, res) => {
-    await getRegions(res);
-  });
-
-  router.get("/getRegionsValuesMap", async (req, res) => {
-    await getRegionsValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getRegionsValuesMap`, async (req, res) => {
-    await getRegionsValuesMap(res);
-  });
-
-  router.get("/getStatistics", async (req, res) => {
-    await getStatistics(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getStatistics`, async (req, res) => {
-    await getStatistics(res);
-  });
-
-  router.get("/getStatisticsValuesMap", async (req, res) => {
-    await getStatisticsValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getStatisticsValuesMap`, async (req, res) => {
-    await getStatisticsValuesMap(res);
-  });
-
-  router.get("/getVariables", async (req, res) => {
-    await getVariables(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getVariables`, async (req, res) => {
-    await getVariables(res);
-  });
-
-  router.get("/getVariablesValuesMap", async (req, res) => {
-    await getVariablesValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getVariablesValuesMap`, async (req, res) => {
-    await getVariablesValuesMap(res);
-  });
-
-  router.get("/getThresholds", async (req, res) => {
-    await getThresholds(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getThresholds`, async (req, res) => {
-    await getThresholds(res);
-  });
-
-  router.get("/getThresholdsValuesMap", async (req, res) => {
-    await getThresholdsValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getThresholdsValuesMap`, async (req, res) => {
-    await getThresholdsValuesMap(res);
-  });
-
-  router.get("/getScales", async (req, res) => {
-    await getScales(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getScales`, async (req, res) => {
-    await getScales(res);
-  });
-
-  router.get("/getScalesValuesMap", async (req, res) => {
-    await getScalesValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getScalesValuesMap`, async (req, res) => {
-    await getScalesValuesMap(res);
-  });
-
-  router.get("/getTruths", async (req, res) => {
-    await getTruths(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getTruths`, async (req, res) => {
-    await getTruths(res);
-  });
-
-  router.get("/getTruthsValuesMap", async (req, res) => {
-    await getTruthsValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getTruthsValuesMap`, async (req, res) => {
-    await getTruthsValuesMap(res);
-  });
-
-  router.get("/getFcstLengths", async (req, res) => {
-    await getFcstLengths(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getFcstLengths`, async (req, res) => {
-    await getFcstLengths(res);
-  });
-
-  router.get("/getFcstTypes", async (req, res) => {
-    await getFcstTypes(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getFcstTypes`, async (req, res) => {
-    await getFcstTypes(res);
-  });
-
-  router.get("/getFcstTypesValuesMap", async (req, res) => {
-    await getFcstTypesValuesMap(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getFcstTypesValuesMap`, async (req, res) => {
-    await getFcstTypesValuesMap(res);
-  });
-
-  router.get("/getValidTimes", async (req, res) => {
-    await getValidTimes(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getValidTimes`, async (req, res) => {
-    await getValidTimes(res);
-  });
-
-  router.get("/getLevels", async (req, res) => {
-    await getLevels(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getLevels`, async (req, res) => {
-    await getLevels(res);
-  });
-
-  router.get("/getDates", async (req, res) => {
-    await getDates(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/getDates`, async (req, res) => {
-    await getDates(res);
-  });
-
-  // users manually enter this URL to refresh the metadata
-  router.get("/refreshMetadata", async (req, res) => {
-    await refreshMetadataMWltData(res);
-  });
-
-  router.get(`${proxyPrefixPath}/:app/refreshMetadata`, async (req, res) => {
-    await refreshMetadataMWltData(res);
-  });
-
-  router.post("/refreshScorecard/:docId", (req, res) => {
-    refreshScorecard(req.params, res);
-  });
-
-  router.post(`${proxyPrefixPath}/:app/refreshScorecard/:docId`, (req, res) => {
-    refreshScorecard(req.params, res);
-  });
-
-  router.post("/setStatusScorecard/:docId", (req, res) => {
-    setStatusScorecard(req.params, req, res);
-  });
-
-  router.post(`${proxyPrefixPath}/:app/setStatusScorecard/:docId`, (req, res) => {
-    setStatusScorecard(req.params, req, res);
-  });
-
-  // mount the router on the app
-  app.use("/", router);
-
-  WebApp.handlers.use(app);
+  // configure the routes
+  setupRouter();
 }
 
 // eslint-disable-next-line no-undef
